+++ ntex/tests/http_client.rs
@@ -706,6 +706,34 @@ async fn client_payload_poll_after_eof() {
 }
 
 #[ntex::test]
+async fn client_early_response() {
+    let addr = ntex::server::TestServer::unused_addr();
+    let lst = std::net::TcpListener::bind(addr).unwrap();
+    std::thread::spawn(move || {
+        for mut stream in lst.incoming().flatten() {
+            let mut b = [0; 1024];
+            let _ = stream.read(&mut b);
+            let _ = stream
+                .write_all(b"HTTP/1.1 413 Payload Too Large\r\ncontent-length: 8\r\n\r\ntoo big!");
+            std::thread::sleep(std::time::Duration::from_millis(200));
+            // close with unread request body
+            drop(stream);
+        }
+    });
+
+    let body = Bytes::from(vec![b'x'; 64 * 1024 * 1024]);
+    let response = Client::builder()
+        .build(ClientConfig::new().set_response_timeout(Seconds(30)))
+        .post(format!("http://{addr}/").as_str())
+        .send_body(body)
+        .await
+        .unwrap();
+    assert_eq!(response.status(), ntex::http::StatusCode::PAYLOAD_TOO_LARGE);
+    let bytes = response.body().await.unwrap();
+    assert_eq!(bytes, Bytes::from_static(b"too big!"));
+}
+
+#[ntex::test]
 async fn client_truncated_body_http10() {
     let addr = raw_server(b"HTTP/1.0 200 OK\r\ncontent-length: 20\r\n\r\nwelcome!");